+++ locals.tf
@@ -3,4 +3,15 @@ locals {
     subnet_ids = var.allowed_subnet_ids
     cidrs      = var.allowed_cidrs
   }
+
+  hubs_auth_rules = flatten([
+    for hub_name, hub in try(var.hubs_parameters, {}) : [
+      for rule in ["listen", "send", "manage"] : {
+        hub_name       = hub_name
+        hub            = hub
+        rule           = rule
+        authorizations = hub.authorizations
+      }
+    ]
+  ])
 }