+++ outputs.tf
@@ -36,21 +36,25 @@ output "namespace_default_authorization_rule_name" {
 output "namespace_default_primary_connection_string" {
   description = "Event Hub namespace default primary connection string"
   value       = azurerm_eventhub_namespace.eventhub.default_primary_connection_string
+  sensitive   = true
 }
 
 output "namespace_default_primary_key" {
   description = "Event Hub namespace default primary key"
   value       = azurerm_eventhub_namespace.eventhub.default_primary_key
+  sensitive   = true
 }
 
 output "namespace_default_secondary_connection_string" {
   description = "Eventhub namespace default secondary connection string"
   value       = azurerm_eventhub_namespace.eventhub.default_secondary_connection_string
+  sensitive   = true
 }
 
 output "namespace_default_secondary_key" {
   description = "Event Hub namespace default secondary key"
   value       = azurerm_eventhub_namespace.eventhub.default_secondary_key
+  sensitive   = true
 }
 
 output "namespace_listen_authorization_rule" {
@@ -64,11 +68,26 @@ output "namespace_send_authorization_rule" {
 }
 
 output "namespace_manage_authorization_rule" {
-  description = "Event Hub namespace manage only authorization rule"
+  description = "Event Hub namespace manage authorization rule"
   value       = try(azurerm_eventhub_namespace_authorization_rule.manage["enabled"], null)
 }
 
 output "consumer_groups" {
   description = "Azure Event Hub Consumer Groups"
   value       = try(azurerm_eventhub_consumer_group.eventhub_consumer_group, null)
+}
+
+output "hubs_listen_authorization_rule" {
+  description = "Event Hubs listen only authorization rules"
+  value       = { for a in local.hubs_auth_rules : a.hub_name => azurerm_eventhub_authorization_rule.listen[a.hub_name] if a.rule == "listen" && a.authorizations.listen }
+}
+
+output "hubs_send_authorization_rule" {
+  description = "Event Hubs send only authorization rules"
+  value       = { for a in local.hubs_auth_rules : a.hub_name => azurerm_eventhub_authorization_rule.send[a.hub_name] if a.rule == "send" && a.authorizations.send }
+}
+
+output "hubs_manage_authorization_rule" {
+  description = "Event Hubs namespace manage authorization rules"
+  value       = { for a in local.hubs_auth_rules : a.hub_name => azurerm_eventhub_authorization_rule.manage[a.hub_name] if a.rule == "manage" && a.authorizations.manage }
 }